optimizing done but some error persists
error: if starting through yarn build and start, force-refreshing from user page occurs an error:

An error occurred in the Server Components render. The specific message is omitted in production builds to avoid leaking sensitive details. A digest property is included on this error instance which may provide additional details about the nature of the error.] {
  digest: '2280890512'

Need to check if production site occurs same error.